libs/cli/pyproject.toml
[project]
name = "cogniverse-cli"
version = "0.1.0"
description = "CLI for deploying and managing Cogniverse"
requires-python = ">=3.12"
dependencies = [
    "click>=8.0",
    "rich>=13.0",
    "httpx>=0.27",
    "pyyaml>=6.0",
]

[project.scripts]
cogniverse = "cogniverse_cli.main:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["cogniverse_cli"]
artifacts = [
    "cogniverse_cli/data/**",
]
